Add extra features to the book rating app.

+++ src/main/kotlin/models/Book.kt
@@ -6,7 +6,7 @@ data class Book(
     val BookGenre: String,
     val BookReleaseYear:Int,
     val BookLength: Int,
-    val BookPrice: Float,
+    val BookPrice: Double,
     val BookIsRecommended: Boolean
      ){
 }
